Changed rhizome version to long instead of float that has resolution problems.  Added exception catching for parseLong that throws uncaught exceptions when, for example, it hits a float.

+++ src/org/servalproject/rhizome/RhizomeFile.java
@@ -283,7 +283,7 @@ public class RhizomeFile {
 	 *            The name of the incoming file
 	 * @param version
 	 */
-	public static void GenerateMetaForFilename(String fileName, float version) {
+	public static void GenerateMetaForFilename(String fileName, long version) {
 		try {
 			Properties metaP = new Properties();
